+++ example/send_gmail.dart
@@ -22,43 +22,42 @@ main(List<String> rawArgs) async {
   }
 
   List<String> tos = args[toArgs] as List<String> ?? [];
-  if (tos.isEmpty)
+  if (tos.isEmpty) {
     tos.add(username.contains('@') ? username : username + '@gmail.com');
+  }
 
-  // If you want to use an arbitrary SMTP server, go with `new SmtpServer()`.
+  // If you want to use an arbitrary SMTP server, go with `SmtpServer()`.
   // The gmail function is just for convenience. There are similar functions for
   // other providers.
   final smtpServer = gmail(username, args.rest[1]);
 
   Iterable<Address> toAd(Iterable<String> addresses) =>
-      (addresses ?? <String>[]).map((a) => new Address(a));
+      (addresses ?? []).map((a) => Address(a));
 
   Iterable<Attachment> toAt(Iterable<String> attachments) =>
-      (attachments ?? <String>[]).map((a) => new FileAttachment(new File(a)));
+      (attachments ?? []).map((a) => FileAttachment(File(a)));
 
   // Create our message.
-  final message = new Message()
-    ..from = new Address('$username@gmail.com', 'My name 😀')
+  final message = Message()
+    ..from = Address('$username@gmail.com', 'My name 😀')
     ..recipients.addAll(toAd(tos))
     ..ccRecipients.addAll(toAd(args[ccArgs] as Iterable<String>))
     ..bccRecipients.addAll(toAd(args[bccArgs] as Iterable<String>))
-    ..subject = 'Test Dart Mailer library :: 😀 :: ${new DateTime.now()}'
+    ..subject = 'Test Dart Mailer library :: 😀 :: ${DateTime.now()}'
     ..text = 'This is the plain text.\nThis is line 2 of the text part.'
     ..html = "<h1>Test</h1>\n<p>Hey! Here's some HTML content</p>"
     ..attachments.addAll(toAt(args[attachArgs] as Iterable<String>));
 
-  final sendReports =
-      await send(message, smtpServer, timeout: new Duration(seconds: 15));
-  sendReports.forEach((sr) {
-    if (sr.sent)
-      print('Message sent');
-    else {
-      print('Message not sent.');
-      for (var p in sr.validationProblems) {
-        print('Problem: ${p.code}: ${p.msg}');
-      }
+  try {
+    final sendReport =
+    await send(message, smtpServer, timeout: Duration(seconds: 15));
+    print('Message sent: ' + sendReport.toString());
+  } on MailerException catch (e) {
+    print('Message not sent.');
+    for (var p in e.problems) {
+      print('Problem: ${p.code}: ${p.msg}');
     }
-  });
+  }
 }
 
 const toArgs = 'to';
@@ -68,7 +67,7 @@ const bccArgs = 'bcc';
 const verboseArg = 'verbose';
 
 ArgResults parseArgs(List<String> rawArgs) {
-  var parser = new ArgParser()
+  var parser = ArgParser()
     ..addFlag('verbose', abbr: 'v', help: 'Display logging output.')
     ..addMultiOption(toArgs,
         abbr: 't',
@@ -87,7 +86,7 @@ ArgResults parseArgs(List<String> rawArgs) {
 
   var attachments = args[attachArgs] as Iterable<String> ?? [];
   for (var f in attachments) {
-    File attachFile = new File(f);
+    File attachFile = File(f);
     if (!attachFile.existsSync()) {
       showUsage(parser, 'Failed to find file to attach: ${attachFile.path}');
       exit(1);