+++ lib/smtp_server/gmail.dart
@@ -1,4 +1,4 @@
 import '../smtp_server.dart';
 
 SmtpServer gmail(String username, String password) =>
-    new SmtpServer('smtp.gmail.com', username: username, password: password);
+    SmtpServer('smtp.gmail.com', username: username, password: password);