+++ lib/smtp_server/qq.dart
@@ -1,4 +1,4 @@
 import '../smtp_server.dart';
 
-SmtpServer qq(String username, String password) => new SmtpServer('smtp.qq.com',
+SmtpServer qq(String username, String password) => SmtpServer('smtp.qq.com',
     ssl: true, port: 465, username: username, password: password);