+++ lib/src/smtp/connection.dart
@@ -3,29 +3,41 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:async/async.dart';
-import 'package:dart2_constant/convert.dart' as convert;
 import 'package:logging/logging.dart';
 import 'package:mailer/smtp_server.dart';
 
+import 'capabilities.dart';
 import 'exceptions.dart';
-import 'server_response.dart';
 
-/**
- * This class contains all relevant data for one smtp connection/session.
- *
- * This includes the socket, smtp-options, but also other objects, relevant
- * for an smtp session, like an input queue.
- **/
+/// This class contains all relevant data for one smtp connection/session.
+///
+/// This includes the socket, smtp-options, but also other objects, relevant
+/// for an smtp session, like an input queue.
 
-final _logger = new Logger('Connection');
+final _logger = Logger('Connection');
+
+class ServerResponse {
+  final String responseCode;
+
+  /// Every line received from the server is one entry in the message list.
+  final List<String> responseLines;
+
+  ServerResponse(this.responseCode, this.responseLines);
+}
 
 class Connection {
-  final SmtpServer _server;
+  final SmtpServer server;
   final Duration timeout;
+  DateTime _connectionOpenStart;
+
+  DateTime get connectionOpenStart => _connectionOpenStart;
+
+  Capabilities capabilities;
+
   Socket _socket;
   StreamQueue<String> _socketIn;
 
-  Connection(this._server, {Duration timeout})
+  Connection(this.server, {Duration timeout})
       : this.timeout = timeout ?? const Duration(seconds: 60);
 
   bool get isSecure => _socket != null && _socket is SecureSocket;
@@ -37,8 +49,8 @@ class Connection {
   /// does not start with any of the strings in [acceptedRespCodes];
   Future<ServerResponse> send(String command,
       {List<String> acceptedRespCodes = const ['2'],
-      String expect: null,
-      bool waitForResponse: true}) async {
+      String expect,
+      bool waitForResponse = true}) async {
     // Send the new command.
     if (command != null) {
       _logger.fine('> $command');
@@ -65,7 +77,7 @@ class Connection {
         (currentLine.length > 3 && currentLine[3] != ' ')) {
       var hasNext = await _socketIn.hasNext.timeout(timeout);
       if (!hasNext) {
-        throw new SmtpClientCommunicationException(
+        throw SmtpClientCommunicationException(
             "Socket was closed even though a response was expected.");
       }
 
@@ -90,10 +102,10 @@ class Connection {
           'After sending $command, response did not start with any of: $acceptedRespCodes.';
       msg += '\nResponse from server: $mString';
       _logger.warning(msg);
-      throw new SmtpClientCommunicationException(msg);
+      throw SmtpClientCommunicationException(msg);
     }
 
-    return new ServerResponse(responseCode, messages);
+    return ServerResponse(responseCode, messages);
   }
 
   /// Upgrades the connection to use TLS.
@@ -102,22 +114,23 @@ class Connection {
     // A StreamQueue always pauses unless we explicitly call next().
     // So we don't need to call pause() ourselves.
     _socket = await SecureSocket.secure(_socket,
-        onBadCertificate: (_) => _server.ignoreBadCertificate);
+        onBadCertificate: (_) => server.ignoreBadCertificate);
     _setSocketIn();
   }
 
   /// Initializes a connection to the given server.
   Future<Null> connect() async {
-    _logger.finer("Connecting to ${_server.host} at port ${_server.port}.");
+    _connectionOpenStart = DateTime.now();
+    _logger.finer("Connecting to ${server.host} at port ${server.port}.");
 
     // Secured connection was demanded by the user.
-    if (_server.ssl) {
-      _socket = await SecureSocket.connect(_server.host, _server.port,
-          onBadCertificate: (_) => _server.ignoreBadCertificate,
+    if (server.ssl) {
+      _socket = await SecureSocket.connect(server.host, server.port,
+          onBadCertificate: (_) => server.ignoreBadCertificate,
           timeout: timeout);
     } else {
       _socket =
-          await Socket.connect(_server.host, _server.port, timeout: timeout);
+          await Socket.connect(server.host, server.port, timeout: timeout);
     }
     _socket.timeout(timeout);
 
@@ -133,16 +146,14 @@ class Connection {
     if (_socketIn != null) {
       _socketIn.cancel();
     }
-    _socketIn = new StreamQueue<String>(_socket
-        .transform(convert.utf8.decoder)
-        .transform(const LineSplitter()));
+    _socketIn = StreamQueue<String>(
+        _socket.transform(utf8.decoder).transform(const LineSplitter()));
   }
 
   void verifySecuredConnection() {
-    if (!_server.allowInsecure && !isSecure) {
+    if (!server.allowInsecure && !isSecure) {
       _socket.close();
-      throw new SmtpUnsecureException(
-          "Aborting because connection is not secure");
+      throw SmtpUnsecureException("Aborting because connection is not secure");
     }
   }
 }