+++ test/send_test.dart
@@ -4,9 +4,8 @@ import 'dart:async';
 import 'package:mailer/mailer.dart';
 import 'package:mailer/smtp_server.dart';
 import 'package:mailer/src/smtp/exceptions.dart';
-import 'package:mailer/src/smtp/smtp_client.dart';
 import "package:test/test.dart";
-import 'dart:convert' as JSON;
+import 'dart:convert';
 
 SmtpServer correctSmtpServer;
 SmtpServer incorrectCredentials = gmail("mister@gmail.com", "wrongpass");
@@ -15,25 +14,22 @@ void main() async {
   correctSmtpServer = await configureCorrectSmtpServer();
 
   test('Sending email', () async {
-
-    List<SendReport> reports = await send(createMessage(correctSmtpServer),
-                                          correctSmtpServer,
-                                          timeout: Duration(seconds: 10));
-    expect(reports.last.sent, true);
+    SendReport report = await send(
+        createMessage(correctSmtpServer), correctSmtpServer,
+        timeout: Duration(seconds: 10));
+    expect(report != null, true);
   }, skip: true);
 
-  test('SmtpClient.checkCredentials() throws SmtpClientAuthenticationException', () async {
-    SmtpClient smtpClient = SmtpClient(incorrectCredentials);
-    expect(smtpClient.checkCredentials(timeout: const Duration(seconds: 5)),
-           throwsA(TypeMatcher<SmtpClientAuthenticationException>()));
-
+  test('SmtpClient.checkCredentials() throws SmtpClientAuthenticationException',
+      () async {
+    expect(checkCredentials(incorrectCredentials, timeout: const Duration(seconds: 5)),
+        throwsA(TypeMatcher<SmtpClientAuthenticationException>()));
   }, skip: false);
-
 }
 
 Future<SmtpServer> configureCorrectSmtpServer() async {
   var config = File('test/smtpserver.json');
-  final json = JSON.jsonDecode(await config.readAsString());
+  final json = jsonDecode(await config.readAsString());
 
   return SmtpServer(
     json['host'] as String,
@@ -47,9 +43,9 @@ Future<SmtpServer> configureCorrectSmtpServer() async {
 
 Message createMessage(SmtpServer smtpServer) {
   // Message to myself
-  return new Message()
-    ..from = new Address(smtpServer.username)
+  return Message()
+    ..from = Address(smtpServer.username)
     ..recipients.add(smtpServer.username)
-    ..subject = 'Test Dart Mailer library :: 😀 :: ${new DateTime.now()}'
+    ..subject = 'Test Dart Mailer library :: 😀 :: ${DateTime.now()}'
     ..text = 'This is the plain text.\nThis is line 2 of the text part.';
 }